Fix pom

+++ hydrator-common/src/test/java/co/cask/hydrator/common/RowRecordTransformerTest.java
@@ -78,7 +78,7 @@ public class RowRecordTransformerTest {
 
     StructuredRecord actual = transformer.toRecord(input);
     Assert.assertTrue((Boolean) actual.get("boolField"));
-    Assert.assertEquals(512L, actual.get("longField"));
+    Assert.assertEquals(512L, (long) actual.get("longField"));
     Assert.assertTrue(Math.abs(3.14f - (Float) actual.get("floatField")) < 0.000001);
     Assert.assertEquals("foo", Bytes.toString((byte[]) actual.get("bytesField")));
     Assert.assertEquals("rock", actual.get("stringField"));
